pyproject.toml
[project]
name = "agent-shield"
version = "0.1.0"
description = "MCP Firewall and Proxy for AI Agents"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]
